login: before creating a custom alias, ensure it does not exist

+++ proot-distro.sh
@@ -202,7 +202,7 @@ command_install() {
 	fi
 
 	if [ -n "${override_alias-}" ]; then
-		if [ ! -e "${DISTRO_PLUGINS_DIR}/${override_alias}.sh" ]; then
+		if [ ! -e "${DISTRO_PLUGINS_DIR}/${override_alias}.sh" ] && [ ! -e "${DISTRO_PLUGINS_DIR}/${override_alias}.override.sh" ]; then
 			msg "${BLUE}[${GREEN}*${BLUE}] ${CYAN}Creating file '${DISTRO_PLUGINS_DIR}/${override_alias}.override.sh'...${RST}"
 			distro_plugin_script="${DISTRO_PLUGINS_DIR}/${override_alias}.override.sh"
 			cp "${DISTRO_PLUGINS_DIR}/${distro_name}.sh" "${distro_plugin_script}"
